refactor : 영화 목록에서 뷰홀더 클래스 분리
